# docs/openapi.yaml
...
openapi: 3.0.0
info:
  title: Udaconnect
  description: Manage persons and their connections
  version: 1.0.0
paths:
...
  /api/persons/{personId}/connection:
    get:
      summary: Get person information by Id
      description: Get person information by id
      operationId: getPersonById
      parameters:
      - name: personId
        in: path
        description: ID of the person to get
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: search results matching criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Person'
        "400":
          description: bad input parameter
...
components:
  schemas:
    Person:
      required:
      - company_name
      - first_name
      - id
      - last_name
      type: object
      properties:
        id:
          type: integer
          example: 1
        first_name:
          type: string
          example: Taco
        last_name:
          type: string
          example: Fargo
        company_name:
          type: string
          example: Alpha Omega Upholstery
...